refactor(custom-hook): moved react custom hooks into common-hooks
