--- apps/web/hooks/useMounted.ts
@@ -1,11 +1,0 @@
-import { useEffect, useState } from 'react';
-
-export const useMounted = () => {
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  return mounted;
-};
