+++ apps/web/libs/suspense/SafeSuspense.tsx
@@ -1,6 +1,6 @@
 import { ComponentProps, Suspense } from 'react';
 
-import { useMounted } from '@hooks/index';
+import { useMounted } from '@common-hooks/index';
 
 export default function SafeSuspense(props: ComponentProps<typeof Suspense>) {
   const isMounted = useMounted();
