+++ apps/web/package.json
@@ -24,7 +24,8 @@
     "react-dom": "18.2.0",
     "react-query": "^3.39.2",
     "ui": "workspace:*",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "common-hooks": "workspace:*"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
