+++ package.json
@@ -24,6 +24,7 @@
     "@changesets/cli": "^2.26.0",
     "@commitlint/cli": "^17.3.0",
     "@commitlint/config-conventional": "^17.3.0",
+    "@types/react": "^18.0.26",
     "commitizen": "^4.2.6",
     "cz-customizable": "^7.0.0",
     "eslint-config-custom": "workspace:0.0.0",
