+++ packages/common-hooks/index.ts
@@ -1,7 +1,6 @@
-export * from '@hooks/useInterval';
+export { useInterval } from './useInterval';
 export { useMounted } from './useMounted';
 export { useForm } from './useForm';
 export { useLocalStorage } from './useLocalStorage';
 export { useSessionStorage } from './useSessionStorage';
-export { useToast } from './useToast';
 export { useWindow } from './useWindow';
