+++ packages/common-hooks/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hooks",
+  "name": "common-hooks",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
