+++ packages/common-hooks/tsconfig.json
@@ -3,7 +3,7 @@
     "baseUrl": ".",
     "paths": {
       "@ui/*": ["../../packages/ui/*"],
-      "@hooks/*": ["./*"]
+      "@common-hooks/*": ["./*"]
     },
     "jsxImportSource": "@emotion/react"
   },
