+++ packages/ui/package.json
@@ -14,7 +14,7 @@
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "eslint": "^8.29.0",
-    "hooks": "workspace:^1.0.0",
+    "common-hooks": "workspace:*",
     "prettier": "^2.8.1",
     "react": "^18.2.0",
     "tsconfig": "workspace:*",
